Reworked updates
Client method update() will now force wt=json and return the decoded
result. Method updateRaw() can be used when this is not desirable.

Added a new property to Update: contentType. Used to denote the content
type of the data in the body.

Added additional options to Update (commit, optimize, overwrite).